www: record raw json report
Modeling each possible piece of cluster data / statistic will become
quite a headache, especially with nested json documents. Doing so is
certainly possible if we have a need for it, but we would do so at the
expense of more complex / slow queries to reconstruct the exact report
from the normalized relational model.

Instead, we stash the raw json report and deal do one-off filtering and
simple queries by manipulating / extracting python dictionary data.

As we get more experience with the types of reports we want, we may want
to go back and deconstruct raw ceph reports.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>