Add a loading spinner to all the settings page
Some of the ajax requests may take some time to complete.
Provide an animation to keep the user aware something is happening.

The font-heights of font-awesome are at odds with the general fonts, so
use an invisible icon to keep the layout from jumping around
when adding and removing the spinner animation via ng-show/ng-hide.