www: migrate space/health to no-history-keeping model
This patch is the beginning of a simplification that doesn't keep a
history of raw dumps. We aren't using them, and they grow quickly,
complicating deployment due to a need for periodic pruning of old
history.

This patch also migrates the conversion of raw data into the view that
the API presents into Kraken in an attemp to decouple as much as
possible from the particular web framework we are using.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>