Remove requirement for build to use global /opt
Making the virtualenvs in $DESTDIR is not trivial; for the 1.0.0
release we just punted and made them in their final location in /opt.
This modifies that process by:

0) building them in the source directory, in 'build' targets;
the install target now just copies from source area to $DESTDIR.
This was necessary because rpm looks for the builddir paths in
all the files, and the .so's still have debug information from
the build directory leaking through.

1) turning off Python byte-compiling so no stray .pyc files get left
with the wrong paths included

2) installing the carbon and graphite-web venvs with the --prefix and
--install-libs options to install into $DESTDIR (why they don't
simply respect $DESTDIR like the rest of the world is unclear)

3) finding all the scripts that hardcode $(DESTDIR)/<venv>/bin/python
and modifying them (with sed) to /opt/... instead

4) also doing the above to the few activate scripts that mentioned
bin/python

5) re-symlinking local/(bin,include,lib) to the /opt (installed) target
(if they exist, as on Ubuntu)

This allows $DESTDIR/opt/{graphite,calamari} to be packaged up for
final installation into /opt/{graphite,calamari} without needing to
use the global directory at build time, which should enable multiple
parallel builds, a very good thing indeed.

Signed-off-by: Dan Mick <dan.mick@inktank.com>