Run mod_wsgi in daemon mode; bizarre bugs result otherwise
The symptom: _convert_to_unix_ms would sometimes be 6 hours off; that
behavior correlated with the 't' datetime.datetime object having no
tzinfo (being 'naive'), and therefore being interpreted as a local time
and the conversion going wacky (against a localtime that was also not
configured).  While arguably Django might run better with no concept of
timezone, still, tracking down the problem proved instructive: by installing
a local pure-Python datetime module, I was able to prove that the datetime
object *did not come from that module*.  Somehow, Apache was polluting
the python interpreter to have loaded a different datetime module and it
was incompatible in weird ways.  An unproven theory is that the 'utc' tzinfo
object that calamari's Django framework provided wasn't recognized as a
subclass of datetime.tzinfo because there were two instances of the
datetime module over time, but who knows?...

Anyway, try to avoid this mess completely by segregating the apps.

Signed-off-by: Dan Mick <dan.mick@inktank.com>