cthulhu: Associate server events with a cluster
This is a for case for when a server is associated with
exactly 1 cluster.  This makes me realise that I should
have made the Event schema support associating
events with multiple FSIDs, but this will cover 99% of cases
for now.

If someone has a server taking part in two clusters,
then the events about that server won't be associated with
either cluster, but will still be visible in the global view
of events.