cthulhu: Fix failure parsing OSD map for pool mapping
The CRUSH rule code was expecting choose_firstn operations
to be choosing direct children, but it should have been
looking for all descendents.

For example, when firstn is used on a hierarchy of
root->hosts->osds, a rule that uses "choose firstn 0 osd"
from the root would get all the osds.

Fixes: #7883