cthulhu: send pings to late minions
This is a bit hacky, but necessary.  The problem is
that when the salt master AES key changes, minions don't
clue in until they receive a message they don't understand.
Our minions are mostly only sending messages, not receiving
them, so don't learn about the new key.  The workaround is
to send a ping to any minions that seem to have gone quiet,
hopefully in time for them to learn about the new key
before generating any spurious "server lost contact"
type alerts.

The alternative would be to do full PKI negotiation on
every event with all the CPU cycles and roundtrips that
that entails.  We may see a better solution in the long
run as salt adds crypto at the transport level.

Because this change makes us functionally dependent on
the expected contact interval, not just for alerting, now
is the time to make the interval robustly defined via
the scheduled interval in the pillar, rather than via
an absolute time in a config setting, so do that to.

Fixes: #7836