cthulhu: avoid generating false reboot events
Only report a reboot if the boot time has
moved forward in time, and if the difference
is substantial (more than 10 seconds).

The boot time that Linux reports can change
for reasons other than rebooting, I'm wondering
if this is just the NTP adjustment after boot
nudging things by about a second here and there.

Fixes: #8093