cthulhu: handle issues with ongoing creations
1. PgCreatingRequest checks pg_num in OSD map
=============================================

A change to the behaviour of PgCreatingRequest
to handle the case where we send ceph a `osd pool set pg_num`
operation, see apparent success, but in fact the OSD map has not
been updated with the correct pg_num (due to a conflict with
some other ongoing pg creation). Now, we check the osd map at
all stages and error out if we don't see the pg_num we expect.

2. Create PoolCreatingRequest
=============================

Previously pool creations completed as soon as the pool as added
to the OSD map.  Now, pool creation requests last until all the
PGs in the pool have left 'creating' state.  This means that callers
doing a "create, modify" sequence can be sure that the pool will
be ready to handle their modifications as soon as the create operation
completes.

Fixes: #8107