rest-api: workaround for salt grain permissions
/var/cache/salt used to be readable from apache, where
calamari_rest used it to read cached minion grains.  That
doesn't work any more as of some revision of salt, so
send the grain reads through the (root) cthulhu process.

This is a hack because it bottlenecks this work through
cthulhu where it should be isolated in one of the pool
of HTTP worker threads.

Signed-off-by: John Spray <john.spray@redhat.com>