Handle OSDs that do not exist
There's a slightly bizarre edge case, where if you have an OSD in DNE
state (i.e. it's in the CRUSH map, but doesn't actually exist), this
causes various errors (IndexError, NotFound etc.) due to the osd_map
being out of sync with both the crush map and service state.  This
commit handles this case in the two places I've so far found problems in
the backend.

There's also one kink in the frontend (it misrepresents the state of a
DNE OSD on the Manage: OSD page, due to an array being out of sync, but
that needs fixing there, not here).

Signed-off-by: Tim Serong <tserong@suse.com>