Add ceph cluster name support for calamari-crush-location
calamari-crush-location.py has a '--cluster' parameter but never makes
use of it, as a result, this script doesn't work when ceph cluster has
a name other than 'ceph'.  This change adds customized ceph cluster
name support for the script so that it doesn't always assume ceph
cluster name to be 'ceph'.

Signed-Off-By: Zhiteng Huang <winston.d@gmail.com>