add %dir directive
The problem is that the rest-api errors out with an IOError: permission
denied as it imports the salt_wrapper and the wrapper tries to intialize
a logHandler for cthulhu.log

This wrapper wants to log import
failures to both calamari.log and cthulhu.log. Because it doesn't know
who is using it. A better fix might be to pass in a logHandler

Without this patch
cthulhu.log will be owned by root as cthulhu runs as root.

The fix is to chown all the log files in /var/log/calamari to the apache
user.

Signed-off-by: Gregory Meno <gmeno@redhat.com>