Cthulhu: Optionally push events to salt event bus
This patch adds an option to push events generated by
calamari to salt event bus. This can be enabled by
setting the option emit_event_to_salt_event_bus to True,
by default this will be set to false. Also there is an
option to set the prefix of the tag of the event pushed.
This can be done using the option "event_tag_prefix". By
default its value is "calamari/". These options are
available in calamari.conf file under [cthulhu].

Signed-off-by: Darshan N <dnarayan@redhat.com>