Packaging: Add SUSE spec file for Calamari-on-MON
This shows how we can do an FHS-path-based install for Calamari
when running on MONs, using distro-packaged dependencies rather
than packing everything into a virtualenv.  Note that calamari-ctl
is not packaged, because AFAICT all it really does now is generate
Django's secret key, so I've put that in the rpm %post script
instead.

Signed-off-by: Tim Serong <tserong@suse.com>