rest-api: change CrushRuleSerializer
The big trade-off here is that inorder to
create a crush rule you need to specify
everything. We can relax the fields required
to create if we implement librados calls to
osd crush rule create-simple and
osd crush rile create-erasure

By going this direction we can include the new
simplified calls later and not need to rev the
API version.

Signed-off-by: Gregory Meno <gmeno@redhat.com>