rest-api: adding views for CRUSH rule
The only method that bears mentioning here
is update, we allow updates of any field
except id which shouldn't be changed
type which I think is a mistake to change
and osd_count which is a field derived from
the osd_map.

Update will fill in any other fields with
the most recent data. We need to do this
because osd setcrushmap won't fill it in for us.

Signed-off-by: Gregory Meno <gmeno@redhat.com>