calamari-common: add a context manager to clean up Rados connections
The problem here is that we were leaking
file descriptors. This wasn't a problem when we
used salt-minion to invoke code using librados
because those processes were short lived.

Now that mon_remote handles the function that we
used to rely on ceph.py for and mon_remote is part
of calamari-lite -- a long running process -- we
need to be better about cleaning up.

Signed-off-by: Gregory Meno <gmeno@redhat.com>