Pass /dev/null as stdin when calling ceph commands
Currently, we run these commands with default stdin which resolves to
tty from which the daemon is run. For commands like 'rbd watch <img>'
which wait for an input, this can result in commands being hung forever.

The patch fixes it by attaching the stdin to /dev/null so the commands
like that will get EOF on stdin and exit immediately.

Signed-off-by: Boris Ranto <branto@redhat.com>