server_monitor: Do not resolve through hostname
If we use getnameinfo to get hostname from an IP and only then we get
fqdn from the hostname then we fail on machines that have several
network interfaces attached each having a different fqdn. In this case,
we get a random fqdn from the IP.

The fix is fairly simple as the socket.getfqdn function does accept an
IP address directly.

Signed-off-by: Boris Ranto <branto@redhat.com>