debian: stop using dh_python2 ${python:Depends}
When dh_python2 parses the `requires.txt` files in Calamari and
Calamari's bundled libraries, it translates the dependencies listed
there into "python-" packages and adds them into ${python:Depends}.

In the past this has led to the calamari-server .deb depending on system
packages that do not exist, which leads to installation failures.

Since ${python:Depends} doesn't work well with Calamari's method of
bundling libraries in a virtualenv, just remove it and hard-code the
list of python Depends in /debian/control.

(I've just copied & pasted the existing list that dh_python2 was already
picking up, so maybe this Depends list could be trimmed further, with
experimentation.)