server_monitor: Register all osd services
If we register managed services only then we can fail when we add
services at a later point through the osd map as the eventer _get_fqdn
requires the osd services to already exist. Otherwise it won't be able
to get fqdn and we get null fqdn for these services.

Signed-off-by: Boris Ranto <branto@redhat.com>