packaging: drop "apache" uid/gid ownership on directories
Now that we run as root and serve HTTPS without Apache, we don't need
to chown directories to the apache user. In fact, since the RPM and Deb
doesn't depend on the Apache packages any more, this UID may not even
exist on the system.

For RPMs, this could lead to UID/GID warnings in
"rpm --verify calamari-server", since the rpmdb thinks directories
should be owned by "apache", but they end up owned by "root" when the
httpd RPM is not installed.